feat: extract a wallet page from ckBTCWallet
