+++ frontend/src/lib/pages/CkBTCWallet.svelte
@@ -1,41 +1,25 @@
 <script lang="ts">
-  import { Island, Spinner } from "@dfinity/gix-components";
-  import Separator from "$lib/components/ui/Separator.svelte";
   import { writable } from "svelte/store";
   import type { WalletStore } from "$lib/types/wallet.context";
   import { debugSelectedAccountStore } from "$lib/derived/debug.derived";
   import { findAccount, hasAccounts } from "$lib/utils/accounts.utils";
   import { icrcAccountsStore } from "$lib/stores/icrc-accounts.store";
-  import { TokenAmount, isNullish, nonNullish } from "@dfinity/utils";
-  import {
-    loadCkBTCAccounts,
-    syncCkBTCAccounts,
-  } from "$lib/services/ckbtc-accounts.services";
-  import { toastsError } from "$lib/stores/toasts.store";
-  import { replacePlaceholders } from "$lib/utils/i18n.utils";
-  import { i18n } from "$lib/stores/i18n";
-  import { goto } from "$app/navigation";
-  import { AppPath } from "$lib/constants/routes.constants";
+  import { isNullish, nonNullish } from "@dfinity/utils";
+  import { loadCkBTCAccounts } from "$lib/services/ckbtc-accounts.services";
   import CkBTCTransactionsList from "$lib/components/accounts/CkBTCTransactionsList.svelte";
   import {
     ckBTCTokenFeeStore,
     ckBTCTokenStore,
   } from "$lib/derived/universes-tokens.derived";
   import CkBTCWalletFooter from "$lib/components/accounts/CkBTCWalletFooter.svelte";
-  import type { UniverseCanisterId } from "$lib/types/universe";
-  import {
-    selectedCkBTCUniverseIdStore,
-    selectedUniverseStore,
-  } from "$lib/derived/selected-universe.derived";
+  import { selectedCkBTCUniverseIdStore } from "$lib/derived/selected-universe.derived";
   import type { CkBTCAdditionalCanisters } from "$lib/types/ckbtc-canisters";
   import { CKBTC_ADDITIONAL_CANISTERS } from "$lib/constants/ckbtc-additional-canister-ids.constants";
   import BitcoinAddress from "$lib/components/accounts/BitcoinAddress.svelte";
   import CkBTCWalletActions from "$lib/components/accounts/CkBTCWalletActions.svelte";
   import type { TokensStoreUniverseData } from "$lib/stores/tokens.store";
   import { loadCkBTCInfo } from "$lib/services/ckbtc-info.services";
-  import IcrcBalancesObserver from "$lib/components/accounts/IcrcBalancesObserver.svelte";
-  import WalletPageHeader from "$lib/components/accounts/WalletPageHeader.svelte";
-  import WalletPageHeading from "$lib/components/accounts/WalletPageHeading.svelte";
+  import WalletPage from "$lib/components/accounts/WalletPage.svelte";
 
   export let accountIdentifier: string | undefined | null = undefined;
 
@@ -47,6 +31,7 @@
   debugSelectedAccountStore(selectedAccountStore);
 
   let transactions: CkBTCTransactionsList;
+  let wallet: WalletPage;
 
   // e.g. is called from "Receive" modal after user click "Done"
   const reloadAccount = async () => {
@@ -55,7 +40,7 @@
     }
 
     await loadCkBTCAccounts({ universeId: $selectedCkBTCUniverseIdStore });
-    await loadAccount($selectedCkBTCUniverseIdStore);
+    await wallet.loadAccount($selectedCkBTCUniverseIdStore);
 
     reloadTransactions();
   };
@@ -66,14 +51,10 @@
     reloadTransactions();
   };
 
-  const reloadOnlyAccountFromStore = () => setSelectedAccount();
-
   // transactions?.reloadTransactions?.() returns a promise.
   // However, the UI displays skeletons while loading and the user can proceed with other operations during this time.
   // That is why we do not need to wait for the promise to resolve here.
   const reloadTransactions = () => transactions?.reloadTransactions?.();
-
-  const goBack = (): Promise<void> => goto(AppPath.Accounts);
 
   const setSelectedAccount = () => {
     selectedAccountStore.set({
@@ -88,74 +69,6 @@
     });
   };
 
-  const loadAccount = async (
-    universeId: UniverseCanisterId
-  ): Promise<{
-    state: "loaded" | "not_found" | "unknown";
-  }> => {
-    setSelectedAccount();
-
-    // We found an account in store for the provided account identifier, all data are set
-    if (nonNullish($selectedAccountStore.account)) {
-      return { state: "loaded" };
-    }
-
-    // Accounts are loaded in store but no account identifier is matching
-    if (hasAccounts($icrcAccountsStore[universeId.toText()]?.accounts ?? [])) {
-      toastsError({
-        labelKey: replacePlaceholders($i18n.error.account_not_found, {
-          $account_identifier: accountIdentifier ?? "",
-        }),
-      });
-
-      await goBack();
-      return { state: "not_found" };
-    }
-
-    return { state: "unknown" };
-  };
-
-  let loaded = false;
-
-  const loadData = async (universeId: UniverseCanisterId | undefined) => {
-    // Universe is not yet loaded
-    if (isNullish(universeId)) {
-      return;
-    }
-
-    // This will display a spinner each time we search and load an account
-    // It will also re-create a new component for the list of transactions which per extension will trigger fetching those
-    loaded = false;
-
-    const { state } = await loadAccount(universeId);
-
-    // The account was loaded or was not found even though accounts are already loaded in store
-    if (state !== "unknown") {
-      loaded = true;
-      return;
-    }
-
-    // Maybe the accounts were just not loaded yet in store, so we try to load the accounts in store
-    await syncCkBTCAccounts({ universeId });
-
-    // And finally try to set the account again
-    await loadAccount(universeId);
-
-    loaded = true;
-  };
-
-  $: accountIdentifier,
-    (async () => await loadData($selectedCkBTCUniverseIdStore))();
-
-  let canMakeTransactions = false;
-  $: canMakeTransactions =
-    nonNullish($selectedCkBTCUniverseIdStore) &&
-    hasAccounts(
-      $icrcAccountsStore[$selectedCkBTCUniverseIdStore.toText()]?.accounts ?? []
-    ) &&
-    nonNullish($ckBTCTokenFeeStore[$selectedCkBTCUniverseIdStore.toText()]) &&
-    nonNullish($ckBTCTokenStore[$selectedCkBTCUniverseIdStore.toText()]);
-
   let canisters: CkBTCAdditionalCanisters | undefined = undefined;
   $: canisters = nonNullish($selectedCkBTCUniverseIdStore)
     ? CKBTC_ADDITIONAL_CANISTERS[$selectedCkBTCUniverseIdStore.toText()]
@@ -166,6 +79,15 @@
     ? $ckBTCTokenStore[$selectedCkBTCUniverseIdStore.toText()]
     : undefined;
 
+  let canMakeTransactions = false;
+  $: canMakeTransactions =
+    nonNullish($selectedCkBTCUniverseIdStore) &&
+    hasAccounts(
+      $icrcAccountsStore[$selectedCkBTCUniverseIdStore.toText()]?.accounts ?? []
+    ) &&
+    nonNullish($ckBTCTokenFeeStore[$selectedCkBTCUniverseIdStore.toText()]) &&
+    nonNullish($ckBTCTokenStore[$selectedCkBTCUniverseIdStore.toText()]);
+
   $: (async () =>
     await loadCkBTCInfo({
       universeId: $selectedCkBTCUniverseIdStore,
@@ -173,72 +95,47 @@
     }))();
 </script>
 
-<Island testId="ckbtc-wallet-component">
-  <main class="legacy" data-tid="ckbtc-wallet">
-    <section>
-      {#if loaded && nonNullish(canisters) && nonNullish($selectedAccountStore.account) && nonNullish($selectedCkBTCUniverseIdStore) && nonNullish(token)}
-        <IcrcBalancesObserver
-          universeId={$selectedCkBTCUniverseIdStore}
-          accounts={[$selectedAccountStore.account]}
-          reload={reloadOnlyAccountFromStore}
-        >
-          <WalletPageHeader
-            universe={$selectedUniverseStore}
-            walletAddress={$selectedAccountStore.account.identifier}
-          />
-          <WalletPageHeading
-            accountName={$selectedAccountStore.account.name ??
-              $i18n.accounts.main}
-            balance={TokenAmount.fromE8s({
-              amount: $selectedAccountStore.account.balanceE8s,
-              token: token?.token,
-            })}
-          >
-            <CkBTCWalletActions
-              reload={reloadAccount}
-              minterCanisterId={canisters.minterCanisterId}
-            />
-          </WalletPageHeading>
+<WalletPage
+  {accountIdentifier}
+  {token}
+  selectedUniverseId={$selectedCkBTCUniverseIdStore}
+  bind:this={wallet}
+>
+  <svelte:fragment slot="header-actions">
+    {#if nonNullish(canisters)}
+      <CkBTCWalletActions
+        reload={reloadAccount}
+        minterCanisterId={canisters.minterCanisterId}
+      />
+    {/if}
+  </svelte:fragment>
 
-          <Separator spacing="none" />
+  <svelte:fragment slot="page-content">
+    {#if nonNullish($selectedAccountStore.account) && nonNullish($selectedCkBTCUniverseIdStore) && nonNullish(canisters)}
+      <BitcoinAddress
+        account={$selectedAccountStore.account}
+        universeId={$selectedCkBTCUniverseIdStore}
+        minterCanisterId={canisters.minterCanisterId}
+        reload={reloadAccount}
+      />
 
-          <!-- Transactions and the explanation go together. -->
-          <div>
-            <BitcoinAddress
-              account={$selectedAccountStore.account}
-              universeId={$selectedCkBTCUniverseIdStore}
-              minterCanisterId={canisters.minterCanisterId}
-              reload={reloadAccount}
-            />
+      <CkBTCTransactionsList
+        bind:this={transactions}
+        account={$selectedAccountStore.account}
+        universeId={$selectedCkBTCUniverseIdStore}
+        indexCanisterId={canisters.indexCanisterId}
+        token={token?.token}
+      />
+    {/if}
+  </svelte:fragment>
 
-            <CkBTCTransactionsList
-              bind:this={transactions}
-              account={$selectedAccountStore.account}
-              universeId={$selectedCkBTCUniverseIdStore}
-              indexCanisterId={canisters.indexCanisterId}
-              token={token?.token}
-            />
-          </div>
-        </IcrcBalancesObserver>
-      {:else}
-        <Spinner />
-      {/if}
-    </section>
-  </main>
-
-  {#if canMakeTransactions}
-    <CkBTCWalletFooter
-      store={selectedAccountStore}
-      {reloadAccount}
-      {reloadAccountFromStore}
-    />
-  {/if}
-</Island>
-
-<style lang="scss">
-  section {
-    display: flex;
-    flex-direction: column;
-    gap: var(--padding-4x);
-  }
-</style>
+  <svelte:fragment slot="footer-actions">
+    {#if canMakeTransactions}
+      <CkBTCWalletFooter
+        store={selectedAccountStore}
+        {reloadAccount}
+        {reloadAccountFromStore}
+      />
+    {/if}
+  </svelte:fragment>
+</WalletPage>
